Remove boost dependency in DataMan tests
This commit removes the dependency that DataMan tests have on Boost's
unit testing framework. This was added in cd54d948f, where boost is
effectively used for unit testing some of the new array changes, but
it's definitely not needed by the DataMan tests, as none of them
includes boost or links against it.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>