Simplify MV* classes definitions
When compiling with gcc 9+, a large number of warnings are produced when
compiling the four classes deriving from MVPosition (i.e., MVBaseline,
MVDirection, MVEarthMagnetic and MVuvw) and their users, with all
warnings indicated that generation of an implicitly-declared copy
constructor for these classes was deprecated.

Implicitly-defined copy constructors are marked as deprecated since
C++11 if a user-provided copy assignment operator or a user-provided
destructor are provided. It turns out these four classes defined both,
although without any clear benefit. This commit removes the definition
of both the copy assignment operator and the user-provided destructor
for these four classes. This allows compilers to generate them for us,
plus the added benefit of getting a copy constructor also defined for
us.

To add to the confusion the header files for these classes allegedly
declared a constructor taking a const MVPosition reference as an
argument, but they were commented as "Copy constructor", giving a false
impression to the reader. In this commit I adjusted the comment of these
constructors to remove the ambiguity, and also updated the class
documentation to mention that construction via const MVPosition
references is (and has always been) also possible.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>