Restore Quantum user-defined copy constructor
The Quantum class' copy semantics are those expected by every C++ value
class: a copy constructor does a deep copy of the original value, and so
does a copy assignment. If the Qtype template parameter obeys the same
semantics, then having these special methods explicitly defined is
therefore unnecessary, as the compiler defines them for us. This was the
rationale behind the changes introduced in 7aeaaf168.

Alas, this reasoning didn't take into account that in some cases Qtype
will not obey these same semantics. In particular, the casacore Array
and Vector classes don't follow this; instead they do copy-by-reference
on copy-construction, and deep-copying on copy assignment. This means
that if one wants to have deep copying on Quantum copy construction a
user-defined constructor needs to be defined (which in turn requires a
user-defined copy assignment operator). This was indeed what was there
before 7aeaaf168, so we are restoring it.

Note that the way Quantum is accommodated to cope with the semantics of
the Array classes is very opaque, as there is no clear link between the
Array classes (and their behavior) and the fact that a user-defined
copy constructor needs to be implemented. To make do I added a big
comment in the constructor body, so future readers are aware of why
things are they are. A better way of doing this would be to specialize
the Quantum class for the Array classes, but that would require more
coding (and testing, maintenance, etc...).

Note also that while this implementation makes Quantum work nicely with
Array classes, there might be other user types with different semantics
(e.g., deep copy on copy constructor, copy by reference on copy
assignment) that will break this implementation.

Finally, since this commit restores the user-defined constructor, it
also restores the user-defined operator. The latter is simply defaulted
though, as it already follows the behavior we want it to follow.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>