Add test around deep-copying of Vectors in Quantum
The semantics of the Quantum class are like any other value C++ class:
on copy construction and copy assignment a deep copy operation should
happen. This behavior was however inadvertently broken in 7aeaaf168, in
part because there was no such test running by default that would check
this behavior (there is tMSMetaData, but it requires extra data and
therefore is skipped automatically by the CI infrastructure).

This commit adds two new checks to the Quantum tests to double-check
that when Quantum objects are copied, their underlying Vector objects
are also deep copied. This will ensure that this behavior is checked
when tests are run, even if tMSMetaData doesn't execute.

These tests would have originally passed, are failing now due to the
changes introduced in 7aeaaf168.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>