Perform explicit cast to signed type
The old version of the code performed an implicit cast from uint64 to
int64. This is probably fine most of the time, but it was based on an
implicit assertion that the number was effectively within the int64
range. This was noted by some compilers (e.g., nvcc) and a warning was
issued.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>