Fix several warnings (#1080)
* Fix several warnings

* Perform explicit cast to signed type

The old version of the code performed an implicit cast from uint64 to
int64. This is probably fine most of the time, but it was based on an
implicit assertion that the number was effectively within the int64
range. This was noted by some compilers (e.g., nvcc) and a warning was
issued.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>

* Remove warnings due to implicit fallthroughs

A number of fallthroughs between switch cases was present in the
codebase, which modern compilers warn about. Warnings can be avoided
though if the [[fallthrough]] attribute, defined from C++17 onwards, is
placed just before the end of the corresponding "case" body. Before 17
there was still support in gcc >= 7 and clang with their own
vendor-specific attributes.

This commit adds a CASACORE_FALLTHROUGH macro definition corresponding
to the correct attribute definition, and sprinkles it across the code
base where implicit fallthroughs were already happening. Most of them
looked genuine; SDPolarizationHandler was the only one that looked like
it needed a break (no pun intended!).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>

* Undo addition of 'final' classes

It would be better to make them final, but I'm not sure this could break downstream software.

* Suppress a few more 'method gets hidden by' warnings

* Solve a few more compiler warnings

* Review comments

Co-authored-by: Rodrigo Tobar <rtobar@icrar.org>