Add support for caching the group boundaries that have been created while sorting the tables.
This avoids calling the sorting functions again when doing a
next() iteration call, since a vector with the iteration
boundaries is cached and hence the information which rows
are part of the next iteration is readily available.

Unit tests for classes Sort and TableIter have been added that
check the new functionality.