Remove unused member from Adios2StMan::impl
The itsStManColumnType member was only ever used when reading/writing
data manager information into the AipsIO stream. The member is not
needed otherwise, and therefore we are removing it. We still read and
write dummy information where itsStManColumnType used to be read/written
to preserve backwards compatibility (probably not a real-world issue as
this storage manager is hardly used, but better to be defensive).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>