Port static fields to user-facing specification
The ADIOS2 storage manager has been relying until now on static fields
to pass certain pieces of information from one instance of the
Adios2StMan object to another. Information passed this way includes the
ADIOS2 I/O engine type, its parameters, and the transports (and their
parameters) attached to the I/O engine. This way of passing information
can be problematic, as it's subject to race conditions, it hides flow of
information, and doesn't adjust overall to how storage managers are
handled by the rest of the CTDS.

This commit improves how the Adios2StMan instances hold and receive/send
this information. Instead of relying on static fields that preserve this
information across different objects, Adios2StMan now:

 * Holds these pieces of information in non-static variables.
 * Exposes its information via well-defined Record fields.
 * Allows this information to be received as a Record object.
 * Clones itself by copying the current fields and feeding a new
   instance with these copies.

With these changes it is now possible to create and configure tables
backed by Adios2StMan from taql and python, which wasn't possible
before. Other common operations available to all storage managers should
now be possible.

Even after these changes, there is still one remaining static member
shared across all instances, which is the MPI communicator. Currently
there is no way to serialise this information into a Record, nor there
is a way to lookup MPI communicators from a central registry given a
key/name, so the situation still needs to be resolved. This has to be
thought a bit more thoroughly though, as the MPI communicator should
probably be handled by the CTDS system to the storage manager rather
than having the storage manager having to track it separately (i.e., the
MPI communicator used by the Table class should be passed).

I've also taken the opportunity to slightly document these now
non-static fields, but documentation is still lacking in most of the
class.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>