Fix FITS-IDI mount types
The mount type codes for orbiting antennas and X-Y mounts were
swapped.  This is probably a historic oddity as AIPS memo 102
that standardized FITS-IDI has the following note:

  VLBA Memo 108 assigned code 2 to X-Y mounts and code 3 to
  orbiting antenna which differs from established usage. The codes
  specified here are consistent with common usage. No standard
  codes have yet been assigned to X-Y mounts.

Codes for X-Y mounts were added later and standardized in AIPS
memo 114 where code 3 is used for X-Y mounts.  This matches
current practice as 3 is used as the mount type for the Hobart
26m antenna in data produced by the LBA (the Australian Long
Baseline Array).