Fix handling of FITS-IDI GAIN_CURVE tables
Fix a pasto that may affect importing GAIN_CURVE tables.  The code was
inspecting the wrang class variable to decide whether the GAIN_CURVE
table should be added to the MeasurementSet.  Fortunately the impact
of this bug was limited since the way the CASA importfitsidi task
uses the code results in the GAIN_CURVE tables getting created anyway.