Make conversion from std::vector to casacore::Vector explicit (#1189)
* Make conversion of std::vector to casacore::Vector explicit

Before this MR, it is possible to implicitly convert a std::vector to a casacore::Vector, for example:

std::vector<int> foo();

casacore::Vector<int> v = foo();

would be fine. Since this can be expensive, this should not be done implicitly. There are several bugs in DP3 because of this, and I also found one casacore type-bug (in RFReaderWriter.cc) by making this conversion explicit.

It might require a few fixes in software that uses casacore, particularly casa (but the upside is that it might also highlight some bugs :) ).

* Review comment by Ger