Remove notice and link (#1185)
* Remove the Link and Notice classes

According to @gervandiepen (in #896) these are unused.
It explicitly changes the behaviour of RecordFieldPtr. Some tests in
Record and RecordField explicitly check for this behaviour, and therefore
fail currently, but all other tests still succeed. It should be checked if this is ok
before merging.

* Remove deprecated List class

This class makes use of Notice so wouldn't compile anymore when removing Notice.

* Fix behaviour of Record & TableRecord

Removing Notice changes the behaviour of Record & TableRecord. These
changes bring it as close as possible to its old behaviour. The
remaining change in behaviour is documented.

* Bring tests in line with new behaviour of (Table)Record

* Clean up

* Remove all deprecated container classes

* Make whatType specialized template to avoid many dummy variables throughout