Skip van Vleck corrections based on new FITS-IDI keyword
As described in the VLBA Sensitivity Upgrade Memo 52, new versions
of the DiFX correlator will apply the van Vleck when requested to do
so.  This will be signalled by a new keyword in the FITS header of
FITS-IDI output.  See

  https://library.nrao.edu/public/memos/vlba/up/VLBASU_52.pdf

for more details.

Recognize this new keyword and skip the van Vleck correction if it
is set to a non-zero value.