Give MVTime explicit comparison operators
The MVTime class didn't have explicit comparison operators except for
the equality operator; however comparisons between objects of this class
could still be done thanks to the automatic conversion from MVTime to
double via MVTime's double conversion operator, which in an "mvtime1
<op> mvtime2" comparison would convert both objects to double values and
perform the operation on the doubles.

The above works well, but only when the MVTime -> double is the only
available implicit conversion. This is usually true, but can break when,
for instance, a new type T has a constructor T(const MVTime &) and the
requested comparison operator "<op>". In this case the original
comparison is ambiguous, as the compiler doesn't know whether to convert
MVTime to double or T, as both "double <op> double" and "T <op> T" are
valid comparisons.

This commit avoid this ambiguity altogether by explicitly adding all
comparison operators to the MVTime class itself. Since the equality
operator was already present this is mostly a copy-paste exercise.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>