Remove duplicate operator implementation
The two stream operators to read/write IPosition objects from/to an
AipsIO object are implemented both here inline and in the
casa/IO/IPositionIO.cc file, which seems like a more natural place for
that code to live, given how they do not have anything to do in
principle with Array objects, nor they have any template parameters (and
hence shouldn't have to be in the ArrayIO.tcc file).

Note that the declaration of these functions still lies in ArrayIO.h.
Moving them over to a new IPositionIO.h file would probably be the best,
but we'd still have to include *that* in ArrayIO.h to avoid breaking
backwards compatibility, so I didn't bother doing it.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>