Allow linking against MPI and ADIOS2 separately
We are aiming for the goal of allowing compiling casacore with ADIOS2
support but not necessarily MPI support. The first step in that
direction is to disentangle both options from the build system, which
were otherwise interwoven. With these changes one now can specify that a
build with ADIOS2 support, but without MPI support, is needed.

Note that ADIOS2 itself might or might not have MPI support. When MPI
support is present in ADIOS, its non-MPI APIs are still available, but
only since version 2.6.0, which is why this is the new required minimum
version.

Note also that some of the existing cmake machinery in tables/ and ms/
was unnecessarily complicated, with them linking explicitly against
MPI/ADIOS2 and adding related compilation flags. The USE_ADIOS2 and
USE_MPI cmake options also permeated too much of the build system, so
I've moved all the logic that depends on these values up to the
top-level CMakeLists.txt file, leaving only variable expansion needed on
the sub-CMakeLists.txt files. Since we are also handling all the
different possibilities already as PUBLIC target links on the casa_table
target there is no need to perform any further adjustments downstream.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>