Add support for ADIOS without MPI on Adios2StMan
The Adios2StMan has historically only supported using ADIOS with MPI,
even though ADIOS itself can be used without MPI support (both for
writing and reading). Even more so, since version 2.6.0 an MPI-enabled
ADIOS can be used in both MPI-enabled or MPI-disabled mode. This
restriction on how we use ADIOS meant we could only ever use it when
compiling the whole of casacore and its Table system with MPI support,
which is not necessarily desired.

This commit brings in the necessary changes to support both MPI-enabled
and -disabled modes of operation for the Adios2StMan. When casacore is
compiled with MPI (USE_MPI=ON) then both modes are made available,
otherwise only the MPI-disabled mode is included.

This new dimension on Adios2StMan construction means we offer now more
constructors to the user. The constructor situation was dire though, as
some parameters have the same type (engineType and configFile are both
strings), and thus it was becoming difficult to support all construction
options. Moreover, with the addition of more MPI-unaware constructors
the choice for the default constructor was less obvious than it used to
be. I have thus changed the constructor list is now structured as
follows:

 * Constructors take a full list of engine type, plus all necessary
   parameters.
 * MPI-aware constructors always need an explicit MPI communicator (no
   more implicit MPI_COMM_WORLD).
 * Construction via config file is done via
   Adios2StMan::from_config_file().
 * The default constructor is the MPI-unaware constructor taking and
   empty engine type and parameters.

There are few experimental applications out there using this storage
manager, so this breakage isn't expected to cause big issues in the
wild.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>