Fix reading FITS files with extra bytes after last HDU
Older FITS-IDI from the VLBA archive have an extra record filled with
zeroes after the last HDU of the file.  This causes the FITS code in
casacore to read the last HDU over and over again.  Fix this by
signalling the end of file even if there are remaining bytes after
the last HDU.