Fix handling of FITS-IDI files with multiple frequency setups
The code that generates the spectral window table only generated
rows for the spectral windows in the first frequency setup.  As
seen with a combined K/Q band VLBA observation from the VLBA
archive, this resulted in references to non-existing data
descriptions in the main table of the measurement set.  Fix this
by generating spectral window rows for all frequency setups.