Fix compilation when using C++20 (#1316)
* Fix compilation when using C++20

C++20 has removed a few deprecated functions, in particular related to allocators. While it might take a while before we want to compile casacore by default with c++20, projects that make use of Casacore and C++20 also fail to compile because casacore headers do not compile, which is probably becoming more standard in the near future.

In this MR I provided a work-around fix to the allocators in Casacore. I should mention though that the way Casacore uses allocators is deprecated, and probably also unnecessary. One of the places where it is used extensively, is in the Block class, which by itself is also easily replaced by something from std, that would also have proper move semantics. So a future task could be te remove all allocator and block functionality, as to simplify the codebase.

* Use allocator_traits to maintain previous behaviour