Correctly set debug/optimized HDF5 libraries
CMake's target_link_libraries allows different libraries to be given
depending on the build type. For this to work, each of the libraries
that need to be specific to a build type need to be preceded in the list
with the build type ("debug" or "optimized"). If a library isn't
preceded by a build type, then it's considered in all build types.

This logic wasn't properly implemented in casacore's FindHDF5 module.
When building the HDF5_LIBRARIES output variable, it tried to build it
combining the debug and optimized libraries it had previously
calculated; however it prepended "debug" only at the start of the full
list of debug libraries, and "optimized" only at the start of the full
list of optimized libraries. This meant that, except the first library
on each list, the rest were considered in all builds.

In practice this has clearly never been an issue. When trying to port
the CI pipeline to run on MacOS 12 though this showed up as the root
cause of runtime problems. The issue in particular was as follows: the
hdf5 homebrew library declares libm as one of its dependencies, so it
shows up in HDF5_LIBRARIES. However, when FindHDF5 is building the
internal HDF5_LIBRARIES_DEBUG list, it looks for "debug"-looking
libraries in the system that match the actual dependencies, using
"{name}d" as a hint. In the case of libm this becomes libmd (a valid
name of a different library for Message Digest funcitons on BSD
systems). In MacOS 12 (at least in the version available in GitHub
Actions) this lookup yields a libmd.tbd file that has no libmd.dynlib
counterpart at runtime. And even though libm*d* should only be a
dependency on debug builds, it still shows up in non-debug builds (the
default) because HDF5_LIBRARIES was incorrectly calculated.

One could argue that this is a platform issue with the MacOS runner, and
one would be right, but on the other hand there was an issue in
casacore's code as well that allows us to bypass this issue altogether,
so it's still valuable to add this fix in.