Switch to building with Ninja
While this is not strictly needed, it helps verifying that we *can*
build with Ninja. Additionally, ninja defaults to using all CPUs, which
isn't the case with make.

Note that this commit not only switches to Ninja, but it also ports the
build execution to use the more generic "cmake --build" command, which
internally calls the required commands depending on the generator, and
doesn't require a change of directory.

I'll also note that the old make invocation was sub-optimal. Using "make
-j" is always a bad idea: contrary to common misconceptions, it spawns
as many processes as possible, without constraining itself to the number
of available CPUs. This can cause not only resource over-subscription
and therefore slower build times, but actual system memory exhaustion,
leading to build failures.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>