Fix docker GitHub workflow
The new 'docker' GitHub workflow used a variable to determine the name of the default branch in the repository.
However, it turned out that this variable can only be used in workflow _templates_.
It is also not possible to use the `github.event` family of variables in the `on:` clause.
So, unfortunately, the only way to currently do this is to use a hard-coded name: 'master'