Modernize the Constants.h file and switch to c++17 (#1346)
* Modernize the Constants.h file and switch to c++17

The Constants.cc fails to compile on my compiler (gcc 13.2), because it complaints that the constants should be declared inside namespace C. Looking at the file
it seems to contain a lot of constants that are declared nowadays in the standard C++ headers. I propose to make the duplicates deprecated.

Also, they can be constexpr which will potentially allow more functions to become constexpr. This however requires C++17. Since modern compilers compile in C++17 by default nowadays, this might be a good moment to switch
(for discussion). It eliminates a .cc file.