Various finishing touches for the compressing storage managers (#1400)
* Various finishing touches for the compressing storage managers

These changes add support for opening a read-only file, and make sure that no writing or truncation is performed when no changes are made. This is e.g. important for supporting MS-es as input in e.g. CWL tasks.

It also adds a magic header mark to the Stokes I data column, such that Stokes I, UVW and Antenna cols can be distinguished from each other already from the header. This changes the format of the Stokes I columns, but since it is not used yet, now is the time (and it should be the last change). This makes the Stokes I stman in line with the other compressing storage managers.

A test set is added for the operations done in the Stokes I stman.

Finally there are some fixes to make sure everything remains robust when failures occur. The RO changes showed that e.g. if a close fails but the state of the file is not updated accurate, the destructor might then also throw, thereby causing a terminate instead of an exception. This should be all fixed now.