Fix thread safety issues in "rc" functions
In order to find the culprit, the functions were also simplified and improved somewhat. The real issue turned out to be to return a reference to a value inside a (Block) container, while that container could be modified by a separate thread.

The thread sanitizer pointer at first to the allocator used by the Block class, which made me suspect that the allocator wasn't thread safe. Hence I replaced it by std::vector, which in the end didn't solve the race issue (it was the returned reference), but which I think is cleaner anyway.