Refactor MCFrame to remove ref to MeasFrame and use pimpl pattern (#1425)
* Refactor MCFrame to remove ref to MeasFrame and use pimpl pattern

This removes an unnecessary MeasFrame field from MCFrame. Since that field keeps a reference, it makes it hard to reason about thread safety.

It also refactors the MCFrame class; the internal fields are moved to the unit file by using the pimpl pattern. This also allows the use of std::optional instead of (smart) pointers, which avoids manual copying (and saves some allocations).

* Fix some comments

* Fix incorrect condition

* Remove unnecessary conditions

* MeasConverters don't need to be optional

* Fix a comment

* Remove unnecessary optionals

* Review comments