Make Sisco work with Dp3 when updating an MS (#1439)
When DP3 updates a column, the chain of calls leads to Casacore calling Sisco while a storage manager file already exists. This causes Sisco to get confused and lets it think a Sisco file exists, but it then crashes when opening the shapes file (which doesn't exist). I think it means that Sisco doesn't quite follow the design that was thought for Casacore storage managers, but also checking if the shapes file exists is a simple fix.