Preserve data manager name in Adios2StMan
The previous implementation returned a hardcoded string that was also
equals to the data manager type name. This both caused confusion and
made it impossible to use two different instances of Adios2StMan in the
same table.

In order to keep changes minimal and maintain backwards compatibility,
the Adios2StMan ctor doesn't contain any changes. Instead a new method
allows to set the data manager name. This is then internally used by
makeObject() and clone().

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>